pyproject.toml
[project]
name = "asyncio-contextvars-pollution"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "httpx",
    "opentelemetry-api",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-sdk",
    "uvicorn",
    "uvloop",
]
